refactor（仿照SpringDoc结构进行改造）
