+++ mybatis-plus-generator-ui-starter-common/src/main/java/com/github/davidfantasy/mybatisplus/generatorui/common/core/properties/SpringDocConfigProperties.java
@@ -1,0 +1,47 @@
+/*
+ *
+ *  *
+ *  *  *
+ *  *  *  * Copyright 2019-2022 the original author or authors.
+ *  *  *  *
+ *  *  *  * Licensed under the Apache License, Version 2.0 (the "License");
+ *  *  *  * you may not use this file except in compliance with the License.
+ *  *  *  * You may obtain a copy of the License at
+ *  *  *  *
+ *  *  *  *      https://www.apache.org/licenses/LICENSE-2.0
+ *  *  *  *
+ *  *  *  * Unless required by applicable law or agreed to in writing, software
+ *  *  *  * distributed under the License is distributed on an "AS IS" BASIS,
+ *  *  *  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  *  *  * See the License for the specific language governing permissions and
+ *  *  *  * limitations under the License.
+ *  *  *
+ *  *
+ *
+ */
+
+package com.github.davidfantasy.mybatisplus.generatorui.common.core.properties;
+
+import com.github.davidfantasy.mybatisplus.generatorui.common.core.utils.Constants;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Lazy;
+import org.springframework.http.MediaType;
+
+import java.util.ArrayList;
+import java.util.List;
+
+/**
+ * The type Spring doc config properties.
+ *
+ * @author bnasslahsen
+ */
+@Lazy(false)
+@Configuration(proxyBeanMethods = false)
+@ConfigurationProperties(prefix = Constants.SPRINGDOC_PREFIX)
+@ConditionalOnProperty(name = Constants.SPRINGDOC_ENABLED, matchIfMissing = true)
+public class SpringDocConfigProperties {
+
+
+}
