+++ mybatis-plus-generator-ui-starter-common/src/main/resources/META-INF/spring.factories
@@ -1,3 +1,3 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
 com.github.davidfantasy.mybatisplus.generatorui.common.core.configuration.SpringDocConfiguration,\
-com.github.davidfantasy.mybatisplus.generatorui.common.core.configuration.SpringDocConfigProperties
+com.github.davidfantasy.mybatisplus.generatorui.common.core.properties.SpringDocConfigProperties
