+++ mybatis-plus-generator-ui-starter-webmvc-ui/pom.xml
@@ -34,5 +34,10 @@
 			<artifactId>tomcat-embed-core</artifactId>
 			<scope>test</scope>
 		</dependency>
-	</dependencies>
+        <dependency>
+            <groupId>org.projectlombok</groupId>
+            <artifactId>lombok</artifactId>
+			<optional>true</optional>
+        </dependency>
+    </dependencies>
 </project>
