+++ mybatis-plus-generator-ui-starter-webmvc-ui/src/main/resources/public/index.html
