+++ mybatis-plus-generator-ui-starter-webmvc-ui/src/main/resources/public/libs/font-awesome/fonts/FontAwesome.otf
