+++ mybatis-plus-generator-ui-starter-webmvc-ui/src/main/resources/public/libs/jquery.fullscreen-min.js
